Temporary install tempest-lib in tox venv
tempest-lib code has been merged back into tempest, but in a different
namespace (tempest.lib). While the old lib is still around, tempest
does not install it anymore, so until a new release of tempest is tagged
so that the existing scripts can be fixed, explicitly install tempest-lib
to allow the old code to run.

Change-Id: I43361680fc55c4b59be0944fbbad4a95b942e7ee