Force installation of tempest-lib on older branches (pre-Mitaka)
Tempest-lib has been reintegrated into tempest (as code, but it still provides
a stable API). The old library is still available, but tempest tests are
failing because when sahara is installed and exports the scenario tests for
python clients, python-lib is not installed (it's correcly in
test-requirements.txt, not requirements.txt), and tempest does not install
tempest-lib anymore.

The tests have been fixed in master (soon Mitaka) but I feel it would be
better to install tempest-lib on the older branches instead of backporting
the patch.

Change-Id: I30154a2d0b8235ca57612fd4421aab5b75b7c2e9