Enable vanilla for tempest tests too
Enable it into Sahara configuration, and configure Tempest for both fake
and vanilla.
This way the API tests in sahara-tests can run.

The long term solution to avoid code duplication is to move the client tests
from sahara to sahara-tests and have them share the base class with the API
tests, so that they can work with any plugin, not just fake.

Change-Id: I938ddc67a3deff6e495ec3be1ebb7e35493cba10